gs_elpa/elpa_db.py: Check version numbers for sanity
Packages with negative version numbers have been seen in melpa-stable.
These appear to be the result of mapping non-numeric versions,
which are defined in version-regexp-alist as follows (Emacs 27.2):
  (("^[-._+ ]?snapshot$" . -4)
   ("^[-._+]$" . -4)
   ("^[-._+ ]?\\(cvs\\|git\\|bzr\\|svn\\|hg\\|darcs\\)$" . -4)
   ("^[-._+ ]?unknown$" . -4)
   ("^[-._+ ]?alpha$" . -3)
   ("^[-._+ ]?beta$" . -2)
   ("^[-._+ ]?\\(pre\\|rc\\)$" . -1))

We could try to map them to Gentoo _alpha, _beta, etc. suffixes, but
it would require more effort to determine the name of the distfile
(and record it in the ebuild). Therefore skip these packages for now.

Signed-off-by: Ulrich Müller <ulm@gentoo.org>